[refactor][FIB-558] move the autolamella widgets into the application's ui package
Eleven widgets under fibsem/ui/widgets/ were named autolamella_* and imported
fibsem.applications.autolamella at runtime. They now live in that application's
own ui package, where their filenames already said they belonged.

This closes the fibsem.ui <-> autolamella.ui runtime import cycle (FIB-555) as a
side effect: autolamella_task_config_editor.py and autolamella_lamella_protocol_editor.py
were the two files reaching back into the application package, and that import is
now intra-package. Nothing under fibsem/ui/ imports fibsem.applications.autolamella.ui
at module level any more.

test_lamella_protocol_editor_canvas.py moves with its widget.
test_spot_burn_coordinates_widget.py stays -- it covers a widget that is not
moving, and only imports the editor inside two functions.

test_report_sections.py reads the report widget as source text through
constructed path segments rather than a module path, so the move broke it in a
way no import-path search would surface. Repointed at the new location.

The destination package's __init__ falls back to AutoLamellaUI = None on
ImportError, so a cycle introduced here would have been swallowed rather than
raised; verified explicitly by importing each moved module first in a fresh
process and asserting AutoLamellaUI still resolves.
