+++ fibsem/applications/autolamella/ui/autolamella_apply_protocol_dialog.py
