+++ fibsem/applications/autolamella/ui/autolamella_generate_report_widget.py
