+++ fibsem/applications/autolamella/ui/autolamella_overview_image_widget.py
