+++ fibsem/applications/autolamella/ui/autolamella_protocol_information_widget.py
